reworked time lib
